CSS-Responsive-Boxes
I struggled a bit with the sizing using VH and VW and just sizing things. I tried multiple variations for each querie with varying degrees of success. I ended up looking at the solution code to check that I was doing it correct and then had to figure out the sizing and positioning doing it the correct way. I still don't feel very confident with the flex ratios.